Solution2

+++ app/main.py
@@ -4,13 +4,14 @@ import os
 def move_file(command: str) -> None:
     if len(command.split()) == 3:
         command_, file_to_move, path_to_move = command.split()
-        if command_ == "mv":
+        if command_ == "mv" and file_to_move != path_to_move.split("/")[-1]:
             place_to_move = path_to_move.split("/")[:-1]
-            for direction in place_to_move:
-                os.mkdir(direction)
-                os.chdir(f"{os.getcwd()}\\{direction}")
-            os.chdir(f"{'../' * len(place_to_move)}/")
+            initial_dir = os.getcwd()
+            for folder in place_to_move:    # Creating folder/folders
+                os.mkdir(folder)
+                os.chdir(f"{os.getcwd()}\\{folder}")
+            os.chdir(initial_dir)   # return to original directory
             with (open(file_to_move, "r") as file_in,
-                  open(os.path.join(path_to_move), "w") as file2):
-                file2.write(file_in.read())
+                  open(os.path.join(path_to_move), "w") as file_out):
+                file_out.write(file_in.read())
             os.remove(file_to_move)
